adicionado funcionalidade listar e ler notas

+++ notes.js
@@ -10,15 +10,25 @@ let addNote = (title, body) => {
         body
     }
 
-    notes.push(note);
+    if(checkIfTitleAlreadyExist(title)){
+        console.log("Titulo da nota já existe, favor alterar Titulo da nova nota.");
+    }else{
+        notes.push(note);
 
-    fs.writeFileSync("notes-data.json", JSON.stringify(notes));
+        fs.writeFileSync("notes-data.json", JSON.stringify(notes));
 
-    console.log("Nota adicionada com sucesso!");
+        console.log("Nota adicionada com sucesso!");
+    }
+    
 };
 
 let getAll = () => {
-    console.log("Pegando todas as notas...");
+    notes.forEach(note => {
+        console.log(`================================
+        \nTítulo: ${note.title}
+        \nConteúdo: ${note.body}
+        \n================================`);
+    });
 };
 
 let removeNote = (title) =>{
@@ -26,14 +36,39 @@ let removeNote = (title) =>{
 };
 
 let getNote = (title) => {
-    console.log("Pegando nota...", title);
+    
+    let resposta = "Nenhuma nota encontrada com este título: \""+title+"\" .";
+    
+    notes.forEach(note =>{
+        if(note.title === title){
+            resposta =`================================
+            \nTítulo: ${note.title}
+            \nConteúdo: ${note.body}
+            \n================================`;
+        }
+    });
+
+    console.log (resposta);
+};
+
+let checkIfTitleAlreadyExist = (title_passado) =>{
+    
+    let exist = false;
+    
+    notes.forEach((note) => {
+        if(title_passado === note.title){
+            exist = true;
+        }
+    });
+
+    return exist;
 };
 
 let loadNotes = () => {
     try{
         notes = JSON.parse(fs.readFileSync('notes-data.json'));
     }catch(e){
-        console.log("ERRO: Arquivo não encontrado!", e);
+        // console.log("ERRO: Arquivo não encontrado!", e);
     }
 };
